+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added default fallback to resolve handler
+
 ## v0.2.10 (2022-11-22)
 * Added normalizer interface structure
 * Renamed Local resolver to Generic